corrected mac number long error

+++ src/main/scala/nano/MAC.scala
@@ -31,10 +31,10 @@ extension (mac: MAC)
     builder += chars(8) += chars(9) += colon
     builder += chars(10) += chars(11)
     builder.toString
-  def number: Long =
+  def number: Int =
     val base = 16
     val chars = mac.toArray
-    var n = 0L
+    var n = 0
     n += Integer.parseInt( s"${chars(0)}${chars(1)}", base )
     n += Integer.parseInt( s"${chars(2)}${chars(3)}", base )
     n += Integer.parseInt( s"${chars(4)}${chars(5)}", base )
